// src/models/enums.ts
export enum Themes {
    DEFAULT = 'DEFAULT',
    DARK = 'DARK',
}

export enum Roles {
    Anonymous = 'Anonymous',
    Student = 'Student',
    Parent = 'Parent',
    Teacher = 'Teacher',
    UnitAdmin = 'UnitAdmin',
    SuperAdmin = 'SuperAdmin',
}

export enum ProjectPageStatus {
    Shared = 'Открытый',
    Private = 'Приватный',
    Banned = 'Забанен'
}